+++ src/gui/backend/chord_selections.c
@@ -75,13 +75,13 @@ chord_selections_paste_to_pos (
     region_timeline_frames_to_local (
       region, playhead->frames, 0);
   position_from_frames (&pos, pos.frames);
-  long pos_ticks = position_to_ticks (&pos);
+  double pos_ticks = position_to_ticks (&pos);
 
   /* get pos of earliest object */
   Position start_pos;
   arranger_selections_get_start_pos (
     (ArrangerSelections *) ts, &start_pos, 0);
-  long start_pos_ticks =
+  double start_pos_ticks =
     position_to_ticks (&start_pos);
 
   /* subtract the start pos from every object and
@@ -91,7 +91,7 @@ chord_selections_paste_to_pos (
   curr_ticks = position_to_ticks (x); \
   position_from_ticks (x, pos_ticks + DIFF)
 
-  long curr_ticks;
+  double curr_ticks;
   int i;
   ChordObject * chord;
   for (i = 0; i < ts->num_chord_objects; i++)
