+++ src/gui/backend/midi_arranger_selections.c
@@ -142,13 +142,13 @@ midi_arranger_selections_paste_to_pos (
     region_timeline_frames_to_local (
       region, playhead->frames, 0);
   position_from_frames (&pos, pos.frames);
-  long pos_ticks = position_to_ticks (&pos);
+  double pos_ticks = position_to_ticks (&pos);
 
   /* get pos of earliest object */
   Position start_pos;
   arranger_selections_get_start_pos (
     (ArrangerSelections *) ts, &start_pos, 0);
-  long start_pos_ticks =
+  double start_pos_ticks =
     position_to_ticks (&start_pos);
 
   /* subtract the start pos from every object and
@@ -158,7 +158,7 @@ midi_arranger_selections_paste_to_pos (
   curr_ticks = position_to_ticks (x); \
   position_from_ticks (x, pos_ticks + DIFF)
 
-  long curr_ticks;
+  double curr_ticks;
   int i;
   for (i = 0; i < ts->num_midi_notes; i++)
     {
